Segregated interpreter from javaspec interaction class
